# SharpAI.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SharpAI.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Image/delete": {
      "delete": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Image/{id}": {
      "delete": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Image/info": {
      "get": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/Image/data": {
      "get": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Image/download": {
      "get": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "jpg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Image/upload-file": {
      "post": {
        "tags": [
          "Image"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/Image/thumbnail-from-data": {
      "get": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "base64ImageData",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pxDiagonal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImageObjData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImageObjInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "bitDepth": {
            "type": "integer",
            "format": "int32"
          },
          "sizeInKb": {
            "type": "number",
            "format": "double"
          },
          "thumbnailBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}